refactor: split create schedule form
